feat: melhorias no layout e separação dos componentes
