+++ src/app/(app)/chamados/components/tickets-dashboard-filters.module.css
@@ -1,0 +1,334 @@
+.filterModalOverlay {
+  position: fixed;
+  inset: 0;
+  z-index: 80;
+  display: flex;
+  align-items: flex-start;
+  justify-content: center;
+  padding: 42px 24px;
+  background: rgba(7, 14, 22, 0.62);
+  backdrop-filter: blur(3px);
+}
+
+.filterModal {
+  width: 100%;
+  max-width: 1220px;
+  max-height: calc(100vh - 84px);
+  overflow: auto;
+  border: 1px solid rgba(74, 107, 138, 0.35);
+  border-radius: 14px;
+  background: linear-gradient(180deg, #102131 0%, #13293c 100%);
+  box-shadow: 0 22px 80px rgba(0, 0, 0, 0.45);
+}
+
+.filterModalHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 18px 18px 8px;
+}
+
+.filterModalTitle {
+  color: #a971ff;
+  font-size: 30px;
+  font-weight: 600;
+  line-height: 1;
+}
+
+.filterModalClose {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  width: 34px;
+  height: 34px;
+  border: none;
+  border-radius: 8px;
+  background: transparent;
+  color: #91a6bc;
+  cursor: pointer;
+}
+
+.filterModalClose:hover {
+  background: rgba(255, 255, 255, 0.05);
+}
+
+.filterModalBody {
+  padding: 8px 16px 12px;
+}
+
+.filterGrid {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 18px 18px;
+}
+
+.filterBlock {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  overflow: visible;
+}
+
+.filterLabel {
+  color: #d9e5f0;
+  font-size: 11px;
+  font-weight: 500;
+  line-height: 1;
+  letter-spacing: 0.04em;
+}
+
+.multiSelectBox {
+  position: relative;
+}
+
+.multiSelectInputWrapper {
+  position: relative;
+}
+
+.multiSelectInput {
+  width: 100%;
+  min-height: 42px;
+  padding: 0 40px 0 14px;
+  border: 1px solid rgba(77, 109, 137, 0.55);
+  border-radius: 8px;
+  background: #0d1c28;
+  color: #d9e7f5;
+  outline: none;
+}
+
+.multiSelectInput:focus {
+  border-color: rgba(116, 154, 189, 0.8);
+}
+
+.multiSelectInput::placeholder {
+  color: #7f98af;
+}
+
+.multiSelectChevron {
+  position: absolute;
+  top: 50%;
+  right: 14px;
+  transform: translateY(-50%);
+  color: #92a9bf;
+  pointer-events: none;
+}
+
+.selectedChips {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 8px;
+  margin-top: 10px;
+}
+
+.selectedChip {
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  min-height: 28px;
+  padding: 0 10px;
+  border-radius: 999px;
+  background: #1b3348;
+  border: 1px solid rgba(98, 128, 154, 0.45);
+  color: #eef4fb;
+  font-size: 12px;
+  font-weight: 500;
+}
+
+.selectedChipRemove {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  border: none;
+  background: transparent;
+  color: #dbe8f5;
+  cursor: pointer;
+  padding: 0;
+}
+
+.multiSelectDropdown {
+  position: absolute;
+  top: calc(100% + 8px);
+  left: 0;
+  right: 0;
+  z-index: 20;
+  min-width: 100%;
+  max-height: 240px;
+  overflow: auto;
+  border: 1px solid rgba(77, 109, 137, 0.55);
+  border-radius: 10px;
+  background: #0d1c28;
+  box-shadow: 0 20px 40px rgba(0, 0, 0, 0.28);
+}
+
+.dropdownHint {
+  padding: 12px 14px;
+  color: #8ea5bb;
+  font-size: 13px;
+}
+
+.dropdownOption {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  width: 100%;
+  min-height: 40px;
+  padding: 0 14px;
+  border: none;
+  background: transparent;
+  color: #e4eef9;
+  cursor: pointer;
+  text-align: left;
+}
+
+.dropdownOption:hover {
+  background: rgba(255, 255, 255, 0.04);
+}
+
+.dateRangePickerWrap {
+  width: 100%;
+}
+
+.dateRangeTrigger {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  width: 100%;
+  min-height: 42px;
+  padding: 0 14px;
+  border: 1px solid rgba(77, 109, 137, 0.55);
+  border-radius: 8px;
+  background: #0d1c28;
+  color: #d9e7f5;
+  font-size: 14px;
+  font-weight: 400;
+  cursor: pointer;
+  outline: none;
+  text-align: left;
+}
+
+.dateRangeTrigger:hover {
+  border-color: rgba(116, 154, 189, 0.8);
+}
+
+.dateRangeTrigger:focus {
+  border-color: rgba(116, 154, 189, 0.8);
+}
+
+.dateRangeTriggerIcon {
+  width: 16px;
+  height: 16px;
+  flex-shrink: 0;
+  opacity: 0.8;
+}
+
+.dateRangeCalendarInline {
+  margin-top: 8px;
+  padding: 12px;
+  background: #0d1c28;
+  border: 1px solid rgba(77, 109, 137, 0.55);
+  border-radius: 8px;
+}
+
+.filterTogglesGrid {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 24px 18px;
+  margin-top: 22px;
+}
+
+.toggleGrid {
+  display: grid;
+  grid-template-columns: repeat(2, minmax(0, 1fr));
+  gap: 8px;
+}
+
+.toggleGridLastFull .toggleButton:last-child {
+  grid-column: 1 / -1;
+}
+
+.toggleButton {
+  min-height: 44px;
+  border-radius: 8px;
+  border: 1px solid #4a5d6d;
+  background: #101d28;
+  color: #f9fafa;
+  font-size: 14px;
+  font-weight: 400;
+  cursor: pointer;
+  transition: background 0.15s ease, border-color 0.15s ease;
+}
+
+.toggleButton:hover {
+  background: #18344d;
+  border-color: #507290;
+}
+
+.toggleButtonActive {
+  border: 1px solid #4a5d6d;
+  background: #1d3449;
+  color: #f9fafa;
+}
+
+.toggleButtonActive:hover {
+  background: #274764;
+  border-color: #4a5d6d;
+}
+
+.filterModalFooter {
+  display: flex;
+  justify-content: flex-end;
+  gap: 12px;
+  padding: 0 16px 16px;
+}
+
+.filterSecondaryButton {
+  min-width: 150px;
+  height: 42px;
+  border: 1px solid rgba(82, 112, 140, 0.5);
+  background: #1d3449;
+  color: #e7f0fa;
+  font-weight: 600;
+}
+
+.filterSecondaryButton:hover {
+  background: #274764;
+}
+
+.filterPrimaryButton {
+  min-width: 150px;
+  height: 42px;
+  border: 1px solid rgba(16, 132, 96, 0.55);
+  background: #0b7a5b;
+  color: #ffffff;
+  font-weight: 700;
+}
+
+.filterPrimaryButton:hover {
+  background: #0d8b68;
+}
+
+@media (max-width: 1100px) {
+  .filterGrid,
+  .filterTogglesGrid {
+    grid-template-columns: 1fr;
+  }
+}
+
+@media (max-width: 640px) {
+  .filterModalOverlay {
+    padding: 16px;
+  }
+
+  .toggleGrid {
+    grid-template-columns: 1fr;
+  }
+
+  .filterModalFooter {
+    flex-direction: column;
+  }
+
+  .filterSecondaryButton,
+  .filterPrimaryButton {
+    width: 100%;
+  }
+}
