+++ src/app/(app)/chamados/criar/page.tsx
@@ -3,9 +3,14 @@ import { TicketCreateHeader } from './components/ticket-create-header'
 
 export default function CriarChamadoPage() {
   return (
-    <div className="page-content space-y-4 overflow-y-scroll pb-24">
-      <TicketCreateHeader />
-      <TicketCreateForm />
+    <div
+      className="page-content flex flex-1 flex-col overflow-y-scroll"
+      style={{ backgroundColor: '#0c161f' }}
+    >
+      <div className="flex flex-1 flex-col gap-8 px-10 pb-10 pt-4">
+        <TicketCreateHeader />
+        <TicketCreateForm />
+      </div>
     </div>
   )
 }
